Move dir. components -> feature/workSession
