+++ features/workSession/components/elements/TabArea/TabComponent/TaskListComponent/CreateTaskButton/CreateTaskButton.tsx
@@ -1,29 +1,30 @@
 import React from 'react';
 import styled from 'styled-components';
-import { ColorThemeName } from '../../../../../types/colorTheme';
+import { Task } from '../../../../../../../../types/entity';
+import { ColorThemeName } from '../../../../../../../../types/colorTheme';
 import {
   astrograniteDebris,
   dryadBark,
   gainsboro,
   washedBlack,
   white,
-} from '../../../../../const/styles/colors';
-import { useAppSelector } from '../../../../../stores/hooks';
-import { selectColorTheme } from '../../../../../stores/slices/colorThemeSlice';
+} from '../../../../../../../../const/styles/colors';
+import { useAppSelector } from '../../../../../../../../stores/hooks';
+import { selectColorTheme } from '../../../../../../../../stores/slices/colorThemeSlice';
 
-type CreateTaskListButtonProps = {
-  onClickCreateTaskList: () => void;
+type CreateTaskButtonProps = {
+  onClickCreateTaskCard: () => void;
 };
-
 const ContainerDiv = styled.div<{
   colorThemeName: ColorThemeName;
 }>`
   position: relative;
   width: 100%;
-  height: 44px;
+  height: 93px;
   display: flex;
-  justify-content: center;
   align-items: center;
+  justify-content: center;
+  padding: 12px;
   border-radius: 8px;
   cursor: pointer;
   &:hover {
@@ -34,6 +35,10 @@ const ContainerDiv = styled.div<{
       if (colorThemeName === 'light') return gainsboro;
       return astrograniteDebris;
     }};
+  color: ${({ colorThemeName }) => {
+    if (colorThemeName === 'light') return dryadBark;
+    return white;
+  }};
   background: ${({ colorThemeName }) => {
     if (colorThemeName === 'light') return white;
     return washedBlack;
@@ -42,29 +47,21 @@ const ContainerDiv = styled.div<{
     colorThemeName === 'light' ? `0 5px 6px 0 ${theme.colors.border}` : 'none'};
 `;
 
-const NameP = styled.p<{
-  colorThemeName: ColorThemeName;
-}>`
-  color: ${({ colorThemeName }) => {
-    if (colorThemeName === 'light') return dryadBark;
-    return white;
-  }};
+const NameP = styled.p`
   font-size: 20px;
   font-weight: 400;
 `;
 
-const CreateTaskListButton = ({
-  onClickCreateTaskList,
-}: CreateTaskListButtonProps) => {
+const CreateTaskButton = ({ onClickCreateTaskCard }: CreateTaskButtonProps) => {
   const currentColorThemeName = useAppSelector(selectColorTheme);
   return (
     <ContainerDiv
       colorThemeName={currentColorThemeName}
-      onClick={onClickCreateTaskList}
+      onClick={onClickCreateTaskCard}
     >
-      <NameP colorThemeName={currentColorThemeName}>+ Create new list</NameP>
+      <NameP>+ Create a task</NameP>
     </ContainerDiv>
   );
 };
 
-export default CreateTaskListButton;
+export default CreateTaskButton;
